refact uri helper move mapbox configure & telemetry to qwant name space
