+++ package.json
@@ -85,13 +85,14 @@
     "ejs": "^2.6.1",
     "express": "^4.16.3",
     "express-static-gzip": "^1.1.3",
-    "mapbox_style_configure": "file:local_modules/mapbox_style_configure",
+    "@qwant/mapbox_style_configure": "file:local_modules/mapbox_style_configure",
     "nconf": "^0.10.0",
     "nconf-yaml": "^1.0.2",
     "nock": "^10.0.0",
     "node-yaml": "3.1.1",
     "prom-client": "^11.1.3",
     "request": "^2.88.0",
-    "telemetry": "file:local_modules/telemetry"
+    "@qwant/telemetry": "file:local_modules/telemetry",
+    "@qwant/uri": "file:local_modules/uri"
   }
 }
